+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
+## Features
+- add option to persist versions to file between startups
+
 ## Changes
 - update `thiserror` dependency  to version 2.x.y
-- update `octocrab` dependency  to version 0.41.y
+- update `octocrab` dependency  to version 0.42.y
 
 # v0.2.0
 ## Changes